refactor(set): extract createSetPriceDto into SetPresenter + presenter spec (W7 remainder)
